Switch to format 3.0 (quilt)
This ensures that there are no spurious changes to upstream sources in
the source package, as dpkg-source aborts early.  Besides, this format
is pretty much standard in Debian these days.

The downside is that any cherry-picks from upstream commits need to be
exported as patches, but I guess we can live with that.  There has
been only one such cherry-pick in the last six years which was already
exported as a patch rather than applied directly.